refactor(spa): extract LinkDetectionSection from TerminalSection
