+++ spa/src/components/settings/TerminalSection.tsx
@@ -5,6 +5,7 @@ import { SettingItem } from './SettingItem'
 import { SegmentControl } from './SegmentControl'
 import { ToggleSwitch } from './ToggleSwitch'
 import { useI18nStore } from '../../stores/useI18nStore'
+import { LinkDetectionSection } from './LinkDetectionSection'
 
 function clamp(value: number, min: number, max: number): number {
   return Math.max(min, Math.min(max, value))
@@ -27,13 +28,6 @@ export function TerminalSection() {
   const setCcIconVariant = useAgentStore((s) => s.setCcIconVariant)
   const showOscTitle = useAgentStore((s) => s.showOscTitle)
   const setShowOscTitle = useAgentStore((s) => s.setShowOscTitle)
-
-  const linkDetectAbsolute = useUISettingsStore((s) => s.linkDetectAbsolute)
-  const setLinkDetectAbsolute = useUISettingsStore((s) => s.setLinkDetectAbsolute)
-  const linkDetectRelativeSlash = useUISettingsStore((s) => s.linkDetectRelativeSlash)
-  const setLinkDetectRelativeSlash = useUISettingsStore((s) => s.setLinkDetectRelativeSlash)
-  const linkDetectBareFilename = useUISettingsStore((s) => s.linkDetectBareFilename)
-  const setLinkDetectBareFilename = useUISettingsStore((s) => s.setLinkDetectBareFilename)
 
   const t = useI18nStore((s) => s.t)
 
@@ -152,20 +146,7 @@ export function TerminalSection() {
         </div>
       </SettingItem>
 
-      <h3 className="text-sm text-text-primary mt-6 mb-1">{t('settings.terminal.link_detect.title')}</h3>
-      <p className="text-xs text-text-secondary mb-3">{t('settings.terminal.link_detect.desc')}</p>
-
-      <SettingItem label={t('settings.terminal.link_detect.absolute.label')} description={t('settings.terminal.link_detect.absolute.desc')}>
-        <ToggleSwitch label={t('settings.terminal.link_detect.absolute.label')} checked={linkDetectAbsolute} onChange={setLinkDetectAbsolute} />
-      </SettingItem>
-
-      <SettingItem label={t('settings.terminal.link_detect.relative_slash.label')} description={t('settings.terminal.link_detect.relative_slash.desc')}>
-        <ToggleSwitch label={t('settings.terminal.link_detect.relative_slash.label')} checked={linkDetectRelativeSlash} onChange={setLinkDetectRelativeSlash} />
-      </SettingItem>
-
-      <SettingItem label={t('settings.terminal.link_detect.bare.label')} description={t('settings.terminal.link_detect.bare.desc')}>
-        <ToggleSwitch label={t('settings.terminal.link_detect.bare.label')} checked={linkDetectBareFilename} onChange={setLinkDetectBareFilename} />
-      </SettingItem>
+      <LinkDetectionSection />
     </div>
   )
 }
